Revert "important change; distinguish serial from waterfall"
ok i'm going to get opinionated here. this is probably one thing
that sets apart async2 from other flow control libraries. really
at its core there are just serial and parallel processes. and all
the derivatives of serial and parallel, such as waterfall, and promises
are really just logic that belongs within the basic serial or
parallel function. because really there are probably tens or hundreds
of ways you could pass data around in a serial or parallel or
combination flow/chain, but not everybody wants a huge library
or museum to download with all those quirky variations. so if you have
a favorite just build it on top of this.

This reverts commit 4706c7c343f16df646628057eaf08a381ee997b4.